Day-4
Started the Data Cleaning process and finalized the IMDb Web scraping.
The code will go to the website and collect the following data about every videogame the website has to offer:
1) Game Title
2) Release date
3) Maturity Rating
4) Genre
5) Director
6) Stars
7) Votes

during the web scraping, the data that isn't available on the website was put as Null (for string columns), and as 0 (for int columns), now that I started Data Cleaning the dataframe, I started working on the "release_date" column to find all the incorrect dates and fix it. I fixed it by having the app webs crap google itself, the app will is in a for loop, it opens a google search page, and types in " video game {game_name} release date" and replace the game_name by the current game that doesn't have a correct release date (the game_name changes every loop), when the search has loaded, it takes the date it finds and replaces it with the correct date relating to the game title in the dataframe.

The next step is to continue this process but with the rest of the columns.