Day-5
Release date is now mostly cleaned, the bot I made limited down the wrong releasedate data from 612 to 289, I did this by automating a bot that google searches for every game's release date and scraps it. I have an idea to use Steam to validate the dates and make sure they are correct and it might as well find some of the remaining missing dates.
Some games unfortunate are just too old or its name is too vague for me to automate it, if I figure out a way to do it I will add it.

What is missing from the project now is to clean the rest of the columns.