Finished Cleaning the release_date
I finished cleaning the release date data. I used Steam to web scrap the release date of the games with wrong date, and also used it to check if the date that was scrapped from scrapping google was correct or not (since using Steam is more reliable in this situation so the date that Steam had override some of the dates that Google found)

Overall, I minimized the wrong data from 612, to 282 using Google, then to 269 using Steam, the steam correction might not sound like much, but remember that it not only corrected the data that was missing, but also checked if the data that already existed was really correct or not.

While I was at it I also deleted all the duplicated rows from the dataframe and organized the CSV files into their own folder with a number showing the newest dataframe.